style(Tab Layouts): Triple Tabs
Changed the tab layout to be on the left side of the screen since we have more scren realestate in the X direction. Added a third tab for "settings" that currently uses the brightness controls we worked out before. Did the math on panel/container sizes so that the 4 panels on the main screen are now equally sized and symmetrically spaced so it looks a bit cleaner.