feat(MPU6050): Bar Graph
Added logic to display angle data calculated from the GY-521 in the symmetrical bar graph. Removed any serial begin stuff to make codebsae cleaner; it can be added back in if debugging. Added a name to one of the charts so I can reference it in code but haven't figured out how to push data to it yet. Gettin there.