feat(Settings Section): Dynamic MPU I2C Address
Created a switch in the settings tab to allow for the MPU address to be dynamically changed between 0x68/0x69. These correspond to pin AD0 not being plugged in/being plugged in to ground, or plugged into power. If I can get it working without setting AD0 high that would be ideal but idk why in tf it keeps fuckin bugging out randomly when nothings connected to that pin. Fuck I'm tired.