feat(Arduino Things): Organization
Moved the backup sketches into their own folder; moved the load cell calibration files also into their own folder. These are both housed in the arduino_things folder so that I know they are separate from the main project, and is not actually part of the ESP32 side of things.