feat(Arduino Code): Updated Motor Control + Load Cell Reading/I2C Sending
Updated the logic handling the operation of the stepper driver to avoid possible blocking while moving the stepper motor inside its own function. The while loop for that now determines which direction it should move, and the function is then just called to move that. Added the logic to allow the arduino to read the load cell values, convert that into newtons, process that into a float, and then transmit that properly formatted value over I2C as a float to the ESP32. This occurs within the request function, which means the ESP32 requests this load cell data, and the arduino then only sends the force data if it is requested.