fix(Arduino Code): Stepper Motor + Load Cell Tuning
Added a missing tare for one of the load cells; lowered the sampling rate from 5 to 1 so that I am not having to wait a whole ass second in between stepper moves. Added logic to disable the driver after 3 seconds of inactivity to avoid the driver from browning out and losing its ability to be controlled.