fix(Arduino Drag Code): Signage
Added a negative in front of the drag variable when setting the drag_newtons variable to send over to the ESP32. Because of the orientation of th load cell, the drag values we expect to see when a force is applied to the test specimen showed up as negative. This corrects that.